refactor: widgets 계층 도입
