+++ src/pages/MyPage/StatisticsPage.tsx
@@ -3,9 +3,9 @@ import SummaryTypeBarChart from "@/features/mypage/ui/SummaryTypeBarChart";
 import TagBubbleChart from "@/features/mypage/ui/TagBubbleChart";
 import TroublogActivityChart from "@/features/mypage/ui/TroublogActivityChart";
 import { useDailyActivityMap } from "@/features/mypage/useDailyActivityMap";
-import { useErrorTagsTop3 } from "@/hooks/useErrorTagsTop3";
-import { useSummaryTypes } from "@/hooks/useSummaryTypes";
-import { useTechTagsTop5 } from "@/hooks/useTechTagsTop5";
+import { useErrorTagsTop3 } from "@/features/mypage/stats/useErrorTagsTop3";
+import { useSummaryTypes } from "@/features/mypage/stats/useSummaryTypes";
+import { useTechTagsTop5 } from "@/features/mypage/stats/useTechTagsTop5";
 
 const StatisticsPage = () => {
   const year = new Date().getFullYear();
