+++ src/widgets/mypage/index.ts
@@ -1,0 +1,2 @@
+export { default as TroubleShootingList } from "./TroubleShootingList";
+export { default as MyPageSidebar } from "./MyPageSidebar";
